# openapi.yaml
openapi: 3.0.0
info:
  title: Automation API
  version: 1.0.0
  description: API for managing members, channels, posts, tasks and analytics
servers:
  - url: /api/v1
    description: Default server
components:
  schemas:
    Member:
      type: object
      required:
        - gmail
        - pwd
      properties:
        _id:
          type: string
          format: objectId
        gmail:
          type: string
          format: email
        pwd:
          type: string
          minLength: 6
        recoveryMail:
          type: string
          format: email
        recoveryMailVerified:
          type: boolean
          default: false
        gmailCreationDate:
          type: string
          format: date-time
        youtubeChannelId:
          type: string
          format: objectId
        systemProfiles:
          type: array
          items:
            type: object
            properties:
              systemName:
                type: string
              profileNumber:
                type: string
        tasks:
          type: array
          items:
            type: string
            format: objectId
...
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              message:
                type: string
paths:
  /members:
    get:
      summary: Get all members
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: sort
          schema:
            type: string
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Member'
    post:
      summary: Create a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Member'
      responses:
        '201':
          description: Member created successfully
  /members/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    get:
      summary: Get member by ID
      responses:
        '200':
          description: Member details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Member'
      responses:
        '200':
          description: Member updated successfully
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete member
      responses:
        '204':
          description: Member deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
...